Overcharge in MSTS brake controller
(cherry picked from commit bf3561ca8b00a2d5777387157821d7ad1a2d6c92)

+++ Source/Orts.Simulation/Simulation/RollingStocks/SubSystems/Brakes/MSTS/AirSinglePipe.cs
@@ -451,7 +451,7 @@ namespace Orts.Simulation.RollingStocks.SubSystems.Brakes.MSTS
                 }
                 if (AuxResPressurePSI > BrakeLine1PressurePSI) // Allow small flow from auxiliary reservoir to brake pipe to restore normal pressure after overcharge
                 {
-                    float dp = elapsedClockSeconds * 0.05f;
+                    float dp = (float)elapsedClockSeconds * 0.05f;
                     if (AuxResPressurePSI - dp < BrakeLine1PressurePSI + dp * AuxBrakeLineVolumeRatio)
                         dp = (AuxResPressurePSI - BrakeLine1PressurePSI) / (1 + AuxBrakeLineVolumeRatio);
                     AuxResPressurePSI -= dp;
